% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStd.R
\name{calcBounds}
\alias{calcBounds}
\title{Calculate Bounds Numerically}
\usage{
calcBounds(FUNmod, par, xrange, seql = 200)
}
\arguments{
\item{FUNmod}{model function}

\item{par}{values of function parameters}

\item{xrange}{range of function domain to be searched for roots}

\item{seql}{length of the search grid}
}
\value{
A named list with upper and lower bounds.
}
\description{
Numerically finds roots for 4th derivative of the model function that are
used to indicate "flat" portions of the curve.
}
\details{
to be added
}
\examples{

}
