% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStd.R
\name{calcBoundsSig}
\alias{calcBoundsSig}
\title{Calculate Bounds for Logistic function}
\usage{
calcBoundsSig(par, xrange)
}
\arguments{
\item{par}{values of function parameters}

\item{xrange}{range of function domain to be searched for roots}
}
\value{
A named list with upper and lower bounds.
}
\description{
Calculates roots for 4th derivative of logistic function that are used to
indicate "flat" portions of the curve.
}
\details{
to be added
}
\examples{

}
