% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{extractStd}
\alias{extractStd}
\title{Extract Standards}
\usage{
extractStd(MFI, stdstr, bgstr, stddil, smpdil, antigen, fname, pdate)
}
\arguments{
\item{MFI}{data frame with variable "Sample" that contains information about
standard concentrations in "1/dilution" format (e.g. "1/200").}

\item{stdstr}{character string indicating standards in the file's "Sample"
column. Not case sensitive. If \code{""} (empty string), standards will be
determined by the pattern "1/" only.}

\item{bgstr}{character string indicating background in the file's "Sample"
column. Not case sensitive.}

\item{stddil}{standard dilutions to use for standard curve fitting. If
\code{NULL}, all the dilutions indicated in a "Sample" variable are used.}

\item{smpdil}{single value for sample dilutions (if dilutions are not
provided in \code{smp} data frame). Ignored if \code{dilvar} is provided
and the variable is included in \code{smp} data frame.}

\item{antigen}{character string.}

\item{fname}{name of the file that contains raw data.}

\item{pdate}{date of the plate processing.}
}
\value{
A list with standards for fitting, background values, sample values,
  indices for samples.
}
\description{
Separates specified serial dilutions for fitting, background, and samples
}
\details{
details
}
