% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeSmp.R
\name{normalizeSmp}
\alias{normalizeSmp}
\title{Normalize Samples}
\usage{
normalizeSmp(smp, smpvar, resvar, dilvar, FUNinv, par, bounds, fitflag,
  fitlog, trim.flat = TRUE)
}
\arguments{
\item{smp}{data frame containing samples}

\item{smpvar}{character string indicating sample variable.}

\item{resvar}{character strings for dilution variable and results.}

\item{dilvar}{character strings for dilution variable and results.}

\item{FUNinv}{inverse function to infer sample concentration}

\item{par}{values of model function parameters}

\item{bounds}{named vector with values for extrema and bounds indicating
"flat" regions of the curve}

\item{fitflag}{flag for the fit as returned by \code{fitStd}}

\item{fitlog}{character string indicating if standard values should be
log-transformed for fitting. If the string contains _x_, \code{xvar} will
be transformed, if it contains _y_ - \code{yvar}.}

\item{trim.flat}{logical value determining how the values of \code{yvar} are
trimmed. If \code{TRUE}, they will be trimmed at the bounds where the curve
starts to flatten out (automatically determined as maxima of the third
derivative of the function). If \code{FALSE}, \code{yvar} will be trimmed
at extrema, defined as the range of standards or asymptotes of the fit
(whichever are less extreme).}
}
\value{
A data frame.
}
\description{
Calculate sample concentration based on the fit of standard dilutions.
}
\details{

}
