% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepLum}
\alias{prepLum}
\title{Reformat raw luminex files for processing}
\usage{
prepLum(antigen, fname, fdir = NULL, dtype = "Median",
  stdstr = "std|stand", bgstr = "blank|background", stddil = NULL,
  smpdil = 1000, nwells = NULL, nsep = 2, ncolmax = 105,
  dformat = "\%m/\%d/\%Y")
}
\arguments{
\item{antigen}{character string.}

\item{fname}{name of the file that contains raw data.}

\item{fdir}{directory where the file is located (alternatively, full path can
be included in \code{fname}).}

\item{dtype}{character string for data type in the file.}

\item{stdstr}{character string indicating standards in the file's "Sample"
column. Not case sensitive. If \code{""} (empty string), standards will be
determined by the pattern "1/" only.}

\item{bgstr}{character string indicating background in the file's "Sample"
column. Not case sensitive.}

\item{stddil}{a vector of standard dilutions. If \code{NULL}, dilutions are
inferred from the file's "Sample" column. Non-null value can be used to
exclude some dilutions from model fitting.}

\item{smpdil}{single value for sample dilutions (if dilutions are not
provided in \code{smp} data frame). Ignored if \code{dilvar} is provided
and the variable is included in \code{smp} data frame.}

\item{nwells}{number of wells. If \code{NULL}, inferred from the file.}

\item{nsep}{number of lines separating different data types in the file.}

\item{ncolmax}{maximum number of columns in the file.}

\item{dformat}{date format in the file.}
}
\value{
A list with standards for fitting, background values, sample values,
  indices for samples.
}
\description{
Read in raw luminex files, extract standards and background (if provided), and format samples for processing.
}
\details{
to be added
}
