% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSmp.R
\name{processSmp}
\alias{processSmp}
\title{Process raw Luminex files}
\usage{
processSmp(
  smp,
  std,
  bg = NULL,
  smpdil = 1,
  fitlog = "xy",
  ismp = 1:nrow(smp),
  plotdir = "./",
  pname = NULL,
  ptitle = "fit and samples",
  xvar = NULL,
  yvar = NULL,
  smpvar = NULL,
  addvar = NULL,
  dilvar = NULL,
  resvar = NULL,
  model = "sigmoid",
  Alow = NULL,
  asym = TRUE,
  trim.flat = TRUE,
  interactive = TRUE,
  monot.prompt = FALSE,
  rm.before = FALSE,
  rm.after = interactive,
  maxrm = 2,
  set.bounds = interactive,
  overwrite.bounds = FALSE,
  ifix = NULL,
  optmethod = "Nelder-Mead",
  maxit = 5000,
  stdcol = c("firebrick3", "darkslategray"),
  rugcol = c("cadetblue", "purple", "firebrick2"),
  width = 7,
  height = 6,
  ...
)
}
\arguments{
\item{smp}{data frame with sample ID, sample values and other optional
variables (e.g. dilution).}

\item{std}{matrix or data frame with standards for fitting.}

\item{bg}{values for background spots.}

\item{smpdil}{single value for sample dilutions (if dilutions are not
provided in \code{smp} data frame). Ignored if \code{dilvar} is provided
and the variable is included in \code{smp} data frame.}

\item{fitlog}{character string indicating if standard values should be
log-transformed for fitting. If the string contains _x_, \code{xvar} will
be transformed, if it contains _y_ - \code{yvar}.}

\item{ismp}{indices for samples to be plotted (e.g. to exclude standards that
are included in \code{smp} data frame).}

\item{plotdir}{directory for the plots to be saved.}

\item{pname}{character string for fit plot. If \code{NULL}, will be based on
\code{ptitle} with underscore substituted for space.}

\item{ptitle}{character string for plot title.}

\item{xvar, yvar}{character strings for the variables used to fit a standard
curve. If \code{NULL}, first two columns are assumed to be \code{x} and
\code{y} variables.}

\item{smpvar}{character string indicating sample variable.}

\item{addvar}{named vector for additional variables (e.g. date or plate),
where element names will be used as variable names in \code{smp}.}

\item{dilvar, resvar}{character strings for dilution variable and results.}

\item{model}{the model to be fit.}

\item{Alow}{lower asymptote for the sigmoid model. If \code{NULL}, the lower
asymptote will be estimated, adding an extra parameter to the model. To fix
the asymptote at the level of background, specify \code{"bg"}. Numeric
value of \code{Alow} will force the asymptote to be fixed at the provided
level.}

\item{asym}{if \code{TRUE}, asymmetry in the fit is allowed, adding an extra
parameter to the model.}

\item{trim.flat}{logical value determining how the values of \code{yvar} are
trimmed. If \code{TRUE}, they will be trimmed at the bounds where the curve
starts to flatten out (automatically determined as maxima of the third
derivative of the function). If \code{FALSE}, \code{yvar} will be trimmed
at extrema, defined as the range of standards or asymptotes of the fit
(whichever are less extreme).}

\item{interactive}{logical value. If \code{TRUE}, the user is prompted to
evaluate the standards (and/or the fit) and asked to remove outliers if
needed. \code{TRUE} value takes precedence over \code{rm.before} and
\code{rm.after}: if both are \code{FALSE}, \code{rm.after} is reset to
\code{TRUE}.}

\item{monot.prompt}{if \code{TRUE}, the user is prompted to evaluate the
standards and possibly remove outliers if the standards are not monotonic
(increasing). \code{FALSE} value is ignored if \code{interactive} is
\code{TRUE}.}

\item{rm.before}{logical value indicating if potential outliers should be
removed before the model is fitted. Ignored if \code{interactive} is
\code{FALSE}.}

\item{rm.after}{logical value indicating if potential outliers should be
removed after the model is fitted. Ignored if \code{interactive} is
\code{FALSE}.}

\item{maxrm}{maximum number of outliers to remove.}

\item{set.bounds}{if \code{TRUE}, the user is given the option to manually
set the bound that is not set automatically. In that case, the prompt
appears even if \code{interactive} is \code{FALSE}.}

\item{overwrite.bounds}{logical value indicating the option to overwrite
automatically set bounds. Ignored if \code{interactive} is \code{FALSE}.}

\item{ifix}{sorted integer vector of length 3 with indices of standards to be
used for getting starting values for optimization.}

\item{optmethod}{method to be used in optimization.}

\item{maxit}{maximum number of iterations in optimization.}

\item{stdcol}{vector of two colors for the standard points and the fit on the
plot.}

\item{rugcol}{vector of three colors for the rugplot, which indicates sample
values (inside the bounds, between the bounds and extrema, and beyond
extrema).}

\item{width, height}{optional parameters for the final saved plot.}

\item{...}{further graphical parameters.}
}
\value{
A list of length three. The first element is a data frame that
  contains the results; the second is a character string with a flag
  containing information about removed points, failure to fit the model,
  manually set bounds, and/or an optional custom note provided by the user
  during an interactive model-fitting procedure. The last element is the
  number of sample values for which the results are trimmed.
}
\description{
Process data for a single antigen: fit a standard curve, establish bounds,
normalize samples, and save a plot showing the fit and the samples.
}
\details{
Note that if \code{fitlog} contains _x_ and thus \code{xvar} in
  standards is log-transform before fitting, the results are still returned
  on a regular scale (values of inverse function are exponentiated and then
  multiplied by dilution).
}
\examples{

}
