% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{read_data}
\alias{read_data}
\title{Read in raw luminex data}
\usage{
read_data(fname, dtype = "Median", nwells = NULL, nsep = 2,
  ncolmax = 105)
}
\arguments{
\item{fname}{name of the file that contains raw data.}

\item{dtype}{character string for data type in the file.}

\item{nwells}{number of wells. If \code{NULL}, inferred from the file.}

\item{nsep}{number of lines separating different data types in the file.}

\item{ncolmax}{maximum number of columns in the file.}
}
\value{
A data frame.
}
\description{
Extract specified data type from the file.
}
\details{
to be added
}
