% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot the fit for standards and the samples}
\usage{
plotFit(std, xvar, yvar, dilvar, fitpar = NULL, FUNmod = NULL,
  iout = NULL, bg = NULL, vsmp = NULL, smpflag = NULL,
  trimval = NULL, trimext = NULL, stdcol = c("firebrick3",
  "darkslategray"), rugcol = c("cadetblue", "purple", "firebrick2"), ...)
}
\arguments{
\item{std}{matrix or data frame with standards for fitting.}

\item{xvar}{character strings for the variables used to fit a standard
curve.}

\item{yvar}{character strings for the variables used to fit a standard
curve.}

\item{dilvar}{character string to check if there is a dilution variable in
\code{std}. If found, used for x-axis labels only.}

\item{fitpar}{values of function parameters.}

\item{FUNmod}{model function.}

\item{iout}{indices of removed standard points.}

\item{bg}{background values.}

\item{vsmp}{sample values.}

\item{smpflag}{character vector, flags for each sample.}

\item{trimval}{for final results, the values at which the samples are
trimmed.}

\item{trimext}{integer vector of length two indicating if the values are
trimmed at the extremum (lower and upper).}

\item{stdcol}{vector of two colors for the standard points and the fit on the
plot.}

\item{rugcol}{vector of three colors for the rugplot, which indicates sample
values (inside the bounds, between the bounds and extrema, and beyond
extrema).}

\item{...}{further graphical parameters.}
}
\description{
Produces a plot that includes points for standards, proposed fit, removed
outliers, bounds for "flat" portions of the curve, and values for samples and
for the background.
}
\details{
to be added
}
